automation/requirements.txt
playwright==1.47.0
pytest==8.3.2
pytest-xdist==3.6.1
allure-pytest==2.13.5
httpx==0.27.2
python-dotenv==1.0.1
black==24.8.0
ruff==0.6.9
pydantic==2.9.2